requirements.txt
yt-dlp
opencv-python 
numpy 
Pillow